rpa_bots/indices_bot/requirements.txt
tagui==1.50.0
pandas==2.2.1
python-dotenv==1.0.1
ruff==0.0.47